+++ internal/provider/storage/storage_volume_resource_alias_test.go
@@ -106,6 +106,14 @@ resource "netapp-ontap_storage_volume_resource" "example" {
     security_style = "mixed"
 	junction_path = "/testacc"
   }
+  autosize = {
+    minimum = 20
+    maximum = 60
+    shrink_threshold = 10
+    grow_threshold = 90
+    mode = "off"
+    size_unit = "mb"
+  }
 }`, host, admin, password, volName, svm)
 }
 
@@ -159,6 +167,14 @@ resource "netapp-ontap_storage_volume_resource" "example" {
     unix_permissions = "755"
     security_style = "mixed"
 	junction_path = "/testacc"
+  }
+  autosize = {
+    minimum = 25
+    maximum = 65
+    shrink_threshold = 15
+    grow_threshold = 95
+    mode = "grow"
+    size_unit = "mb"
   }
 }`, host, admin, password, volName, svm)
 }